Finish background rendering functions

+++ emu/core/oam.cpp
@@ -26,7 +26,7 @@ void PPU::OAM::write(uint16_t addr, uint8_t data)
 
 }
 
-uint24 PPU::output_sppixel()
+uint8_t PPU::output_sppixel()
 {
     return 0;
 }
